requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==27.0.0
importlib-metadata==8.4.0
Markdown==3.7
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
zipp==3.20.1
